Implemented Sword Beam + "ProjectileAttack" script
Implemented a "Projectile Attack" script that summons projectiles using object pooling.
It has several serialized fields to make it adjustable and versatile for future attacks:
- Spawn Position: Where the projectile spawns relative to the player
- Velocity: How fast the projectile tavels. Has both X and Y values.
- Gravity: If it's 0, the projectile isn't affected by gravity. Any other value determines how fast the projectile will fall.
- Projectile Lifetime: Decides how long the projectile is active for.
- Collides with Walls: Determines if the projectile can or can't go through walls
- Sprite: Determines the sprite for the projectile
- Projectile Size: Determines how large the projectile is.

Using this script, I have implemented an attack for Gary, "Sword Beam", triggered when pressing "P". It behaves how I described it in the Discord.